ghaccount/requirements.txt
alembic==1.4.3
blinker==1.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
openpyxl==3.0.5
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
WTForms==2.3.3